# Use `--depth 1` to speed up publish.sh